[Scripting] Use Number as a return value for BucketAggregationScript (#35653) (#35919)
This change fixes #35351. Users were no longer able to return types of 
numbers other than doubles for bucket aggregation scripts. This change 
reverts to the previous behavior of being able to return any type of number 
and having it converted to a double outside of the script.

+++ modules/lang-expression/src/main/java/org/elasticsearch/script/expression/ExpressionScriptEngine.java
@@ -118,7 +118,7 @@ public class ExpressionScriptEngine extends AbstractComponent implements ScriptE
             BucketAggregationSelectorScript.Factory wrappedFactory = parameters -> new BucketAggregationSelectorScript(parameters) {
                 @Override
                 public boolean execute() {
-                    return factory.newInstance(getParams()).execute() == 1.0;
+                    return factory.newInstance(getParams()).execute().doubleValue() == 1.0;
                 }
             };
             return context.factoryClazz.cast(wrappedFactory);
